Move vocabulary processor to examples
