+++ docs/code/processors.rst
@@ -46,13 +46,3 @@ Task Processors
 ----------------------------------
 .. autoclass:: forte.processors.srl_predictor.SRLPredictor
     :members:
-
-:hidden:`Alphabet`
-----------------------------------
-.. autoclass:: forte.processors.vocabulary_processor.Alphabet
-    :members:
-
-:hidden:`VocabularyProcessor`
-----------------------------------
-.. autoclass:: forte.processors.vocabulary_processor.VocabularyProcessor
-    :members:
